==> 00_Initial_Data_Cleaning.ipynb <==
## Imports

import numpy as np
import pandas as pd
import re
import json
import time
import datetime

## Data

%%time
data = "dataset/Cell_Phones_and_Accessories_5.json.gz"
df_phones = pd.read_json(data, lines = True, compression = "gzip")
df_phones.head()

### Load other datasets

%%time
data = "dataset/Video_Games_5.json.gz"
df_videogames = pd.read_json(data, lines = True, compression = "gzip")
df_videogames.head()

%%time
data = "dataset/Automotive_5.json.gz"
df_automotive = pd.read_json(data, lines = True, compression = "gzip")
df_automotive.head()

%%time
data = "dataset/Grocery_and_Gourmet_Food_5.json.gz"
df_grocery = pd.read_json(data, lines = True, compression = "gzip")
df_grocery.head()

## Combine Dataframes

df_videogames.shape

df_phones.shape

df_automotive.shape

df_grocery.shape

df = df_videogames.append(df_phones)
df = df.append(df_automotive)
df = df.append(df_grocery)
df.head()

df.shape

Remove rows with missing values from `reviewText`:

df = df[df.reviewText.notnull()].reindex()

Remove rows associated with non-verified reviews

df = df[df.verified].reindex()

df.shape

Definition of `X` and `y` (class to predict).

X = df.reviewText.values
y = df.overall.values > 3

## Preprocessing/Normalization

#### Emoticons Dictionary

emoticon_repl = {
    # positive emoticons
    r":-?d+": " good ", # :D, :-D, :DD, ecc. in lower case
    r":[- ]?\)+": " good ", # :-), :), :-)), :)), ecc
    r";-?\)+": " good ", # ;). ;)). ;-), ecc.
    r"\(+-?:": " good ", # (:, (-:, ecc.
    r"=\)+" : " good ",
    r"\b<3\b" : " good ",    
    # negative emoticons
    r"[\s\r\t\n]+:/+": " bad ", # :/
    r":\\+": " bad ", # :\
    r"[\s\r\t\n]+\)-?:": " bad ",  # ):, )):, )-:, ecc.
    r":-?\(+": " bad ", # :(, :-(, :((, ecc.
    r"[\s\t\r\n]+d+-?:": " bad "
}

We have used the following function to check if an emoticon is present in some review:

# Check if re is present in some review
def check_presence(regex, stop = 1, verbose = True):
    j = 0
    tv = False
    for i in range(len(X)):
        if re.search(regex, X[i].lower()):
            tv = True
            j += 1
            sentiment = "positive" if y[i] else "negative"
            if verbose:
                print("Review #",i)
                print("Sentiment: {}".format(sentiment))
                print("Occurrences of the regex {} :".format(regex), re.findall(regex, X[i].lower()))
                print("Review text:", X[i]+"\n")
            if j == stop:
                break
    return tv

check_presence(r":-?d+")

check_presence(r"[\s\r\t\n]+\)-?:")

Function for checking if an emoticon corresponds to the expected sentiment:

from scipy.stats import norm
from scipy.stats import chi2_contingency

def check_sentiment(regex):
    pos = 0
    neg = 0
    lx = len(X)
    
    for i in range(lx):
        if re.search(regex, X[i].lower()):
            if y[i]:
                pos += 1
            else:
                neg += 1
                
    tot_reviews = len(y)
    tot_pos = np.sum(y)
    tot_neg = tot_reviews - tot_pos
                
    n1 = pos + neg
    n2 = tot_reviews - n1
    
    if emoticon_repl[regex] in [" good ", " great "]:
        test1 = pos
        test2 = tot_pos - pos
        sent = "positive"        
    else:
        test1 = neg
        test2 = tot_neg - neg
        sent = "negative"

        
    # One tailed proportion test
    # https://online.stat.psu.edu/stat414/node/268/
    p1 = test1/n1
    p2 = test2/n2
    p = (test1 + test2)/(n1+n2)
    num = (p1-p2)
    denom = np.sqrt(p*(1-p)*(1/n1 + 1/n2))
    prop_zstat = num/denom
    prop_pvalue = norm.sf(abs(prop_zstat))
    
    
    # Chisq test
    obs = np.array([[pos, neg],
                    [(tot_pos-pos), (tot_neg-neg)]])
    chi_zstat, chi_pvalue, _, _ = chi2_contingency(obs)
    
    cont_table = "\n{0:>16s} | No {0}\nPos {1:12d} | {2:d}\nNeg {3:12d} | {4:d}\n"\
    .format(regex, pos,tot_pos-pos, neg, tot_neg-neg)
    
    #"Positive reviews containing regex {} : {}; p1 = {:.2f}\n".format(regex, pos, pos/n1) + \
    #"Positive reviews not containing regex {} : {}; p2 = {:.2f}\n".format(regex, tot_pos-pos, (tot_pos-pos)/n2) + \
    print("Contingency table:\n" + \
          cont_table + \
          "\np1: #({} reviews) / #(reviews containing {}) = {:.2f}\n".format(sent, regex, test1/n1) + \
          "p2: #({} reviews) / #(reviews not containing {}) = {:.2f}\n".format(sent, regex, test2/n2) + \
          "H_0 proportion test: p1 - p2 <= 0\n" + \
          "Proportion test p-value = {}\n".format(prop_pvalue) + \
          
          "Chisq test p-value = {}".format(chi_pvalue))

For instance, we would expect that the proportion of positive reviews containing ":)" and similar is larger than the proportion of positive reviews that do not contain ":)".

Similarly, we would expect that the proportion of negative reviews containing ":(" and similar is larger than the proportion of negative reviews that do not contain ":(".

In order to check this, we test the significance of this difference for each emoticon in the dictionary:

# Commented out for slowness and long output, interesting cases follow
# for k in emoticon_repl:
#     print(k)
#     # check_sentiment_chisq(k)
#     check_sentiment(k)
#     print("\n")

Example:

check_sentiment(r":[- ]?\)+")

There are more negative reviews containing "):" and similar than positive ones, even though the difference is not that big:

check_sentiment(r"[\s\r\t\n]+\)-?:")

The emoticon is very rare, so we delete it from the list:

del emoticon_repl[r"[\s\r\t\n]+\)-?:"]

The "D:" is very rare, and the difference in proportions is not significant.

check_sentiment(r"[\s\t\r\n]+d+-?:")

del emoticon_repl[r"[\s\t\r\n]+d+-?:"]

#### Slang and abbreviations/acronyms Dictionary

Old approach: download list of internet acronyms and slang from http://www.netlingo.com/acronyms.php. Resulted in too many terms, with many corresponding to common words.

Manual definition of a dictionary:

slang_repl = {
    "tbh" : "to be honest", "afaik" : "as far as i know",
    "ama" : "ask me anything", "b4" : "before",
    r"baa+d" : " bad ", "brb" : "be right back",
    "btaim" : "be that as it may", "bts" : "behind the scenes",
    "btw" : "by the way", "dyk" : "did you know",
    "eli5" : "explain like i am five", "fomo" : "fear of missing out",
    "ftfy" : "fixed that for you", "ftw" : "for the win",
    "fyi" : "for your information", "g2g" : "got to go",
    r"gooo+d": " good ", "gtg" : "got to go",
    "gg" : "good game", r"gr8t*": " great ",
    "gtr" : "got to run", "hmb" : "hit me back",
    "hmu" : "hit me up", "hth" : "happy to help",
    "icymi" : "in case you missed it", "idc" : "i do not care",
    "idk" : "i do not know", "ikr" : "i know right",
    "ily" : "i love you", "imho" : "in my humble opinion",
    "imo" : "in my opinion", "irl" : "in real life",
    "jk" : "just kidding", "lmao" : "laughing my ass off",
    "lmk" : "let me know", "lol" : "laughing out loud",
    "mfw" : "my face when", "nbd" : "no big deal",
    "nm" : "not much", "nvm" : "never mind",
    "omw" : "on my way", "op" : "original poster",
    "ppl" : "people", "rofl" : "rolling on the floor laughing",
    "roflmao" : "rolling on the floor laughing my ass off", "smh" : "shaking my head",
    "tbbh" : "to be brutally honest", "tfw" : "that feeling when",
    "til" : "today i learned", "tmi" : "too much information",
    "wbu" : "what about you", "yolo" : "you only live once"  
}

# Keep only slang that is present in the reviews
# from tqdm.notebook import tqdm
# 
# slang_keys = list(slang_repl.keys())
# 
# present_slang = [k for k in tqdm(slang_keys) if check_presence(r"\b"+k+r"\b", verbose = False)]

# slang_repl_subset = {k : slang_repl[k] for k in present_slang}

# with open('slang_subset_manual.json', 'w') as fid:
#     json.dump(slang_repl_subset, fid)

with open('dataset/slang_subset_manual.json', 'r') as fid:
    slang_repl_subset = json.load(fid)

set(slang_repl.keys())-set(slang_repl_subset.keys())

slang_repl = slang_repl_subset

#### Contracted forms

We will be interested in combinations of "not" + other terms and similar.

contracted_repl = {
    # casi particolari
    r"won\'t" : "will not", r"won\'" : "will not",
    r"can\'t": "can not", r"shan\'t": "shall not",
    r"shan\'": "shall not", r"ain\'t": "is not",
    r"ain\'": "is not",
    # casi generali
    r"n\'t": " not",
    r"n\'": " not",
    r"\'s": " is",
    r"\'ve": " have", 
    r"\'re": " are", 
    r"\'ll": " will", # Might also be "shall", in any case both will be considered stop words
    r"\'d": " would", # Might also be "had", in any case both will be considered stop words
}

### Complete preprocessing function

Complete preprocessing function, including:

- Lowering the text;
- Removing video html part from reviews including a video;
- Removing URLs;
- Substituting emoticons with associated sentiment;
- Translating slang and acronyms;
- Expanding contracted forms;
- Deleting non-alphanumeric characters (excluding the ones useful for tokenization);
- Deleting terms consisting of digits only.

Stop words removal and spelling correction are done after tokenization.

def preprocess(sent, translate_slang = True):
    
    sent = sent.lower()
    sent = re.sub(r'^<div id="video.*>&nbsp;', '', sent) # Video-review part
    sent = re.sub('https?://[A-Za-z0-9./]+', '', sent) # URLs
    
    for k in emoticon_repl:
        sent = re.sub(k, emoticon_repl[k], sent)

    if translate_slang:
        for k in slang_repl:
            sent = re.sub(r"\b"+re.escape(k)+r"\b", slang_repl[k], sent)
            # sent = sent.replace(k, slang_repl[k])
        
    for k in contracted_repl:
        sent = re.sub(k, contracted_repl[k], sent)
    
    sent = re.sub('[/]+', ' ', sent) # word1/word2 to word1 word2
    # Remove non-alphanumeric characters (but not - and _, might be useful for tokenization)
    sent = re.sub('[^A-Za-z0-9-_ ]+', '', sent)

    # Remove words that are digits only
    sent = re.sub('\b\d+\b', '', sent)
    
    return sent

Example:

X[1272]

preprocess(X[1272])

# X[1697]
X[27]

preprocess(X[27])

%%time
df["reviewTextPreprocessed"] = df["reviewText"].apply(preprocess)

df["reviewText"].loc[1272]

df["reviewTextPreprocessed"].loc[1272]

Delete rows where reviews are empty after preprocessing them:

df = df[~df.reviewTextPreprocessed.str.contains(r"^\s*$")]

98 reviews are empty after preprocessing:

df.shape

len(X)-len(df)

Save cleaned dataset:

df[["reviewTextPreprocessed", "overall", "verified"]].to_csv("dataset/preprocessed_dataset.csv",
                                                             index = False)
