==> 03_Tweets_Sentiment_Analysis.ipynb <==
## Imports

import pandas as pd
import json
import re
import numpy as np
import nltk
import itertools
from sklearn.metrics import precision_recall_curve, auc, confusion_matrix
from sklearn.metrics import classification_report, f1_score, fbeta_score, precision_score, recall_score
from sklearn.feature_extraction.text import TfidfVectorizer
from joblib import dump, load

class StemmedTfidfVectorizer(TfidfVectorizer):
    def build_analyzer(self):
        analyzer = super(TfidfVectorizer, self).build_analyzer()
        return lambda doc: english_stemmer.stemWords(analyzer(doc))

import Stemmer
english_stemmer = Stemmer.Stemmer('en')

## Tweet Data

SANTweets = pd.read_csv("dataset/SandersAnalyticsTweets.csv")
SANTweets["text"] = SANTweets["TweetText"]
SANTweets = SANTweets[["Topic", "Sentiment", "text"]]
SANTweets.head()

SANTweets.shape

SANTweets.groupby("Sentiment").size()

SANTweets.groupby("Topic").size()

def translate_sent(s):
    if s == "irrelevant" or s == "neutral":
        s = 0
    if s == "positive":
        s = 1
    if s == "negative":
        s = -1
    return s

SANTweets["sent"] = SANTweets["Sentiment"].apply(translate_sent)

SANTweets.head()

SMTweets = pd.read_csv("dataset/SamsungTweetsSent.csv")
SMTweets.head()

SMTweets.groupby("sent").size()

CESTweets = pd.read_csv("dataset/ces2020_tweets_full_text.csv")
CESTweets.head()

## Cleaning and Preprocessing Tweets

Dictionaries and preprocessing function from previous notebook:

emoticon_repl = {
    # positive emoticons
    r":-?d+": " good ", r":[- ]?\)+": " good ", r";-?\)+": " good ",
    r"\(+-?:": " good ", r"=\)+" : " good ", r"\b<3\b" : " good ",    
    # negative emoticons
    r"[\s\r\t\n]+:/+": " bad ", r":\\+": " bad ", r"[\s\r\t\n]+\)-?:": " bad ",
    r":-?\(+": " bad ", r"[\s\t\r\n]+d+-?:": " bad "
}

contracted_repl = {
    # casi particolari
    r"won\'t" : "will not", r"won\'" : "will not", r"can\'t": "can not", r"shan\'t": "shall not",
    r"shan\'": "shall not", r"ain\'t": "is not", r"ain\'": "is not",
    # casi generali
    r"n\'t": " not", r"\'t": " not", r"n\'": " not", r"\'s": " is", r"\'ve": " have", 
    r"\'re": " are", 
    r"\'ll": " will", r"\'d": " would",
}

with open('dataset/slang_subset_manual.json', 'r') as fid:
    slang_repl = json.load(fid)
    
def preprocess(sent, translate_slang = True):
    
    sent = sent.lower()
    sent = re.sub(r'^<div id="video.*>&nbsp;', '', sent) # Video-review part
    sent = re.sub('https?://[A-Za-z0-9./]+', '', sent) # URLs
    
    for k in emoticon_repl:
        sent = re.sub(k, emoticon_repl[k], sent)

    if translate_slang:
        for k in slang_repl:
            sent = re.sub(r"\b"+k+r"\b", slang_repl[k], sent)
        
    for k in contracted_repl:
        sent = re.sub(k, contracted_repl[k], sent)
    
    sent = re.sub('[/]+', ' ', sent) # word1/word2 to word1 word2
    sent = re.sub('[^A-Za-z0-9-_ ]+', '', sent)
    sent = re.sub('\b\d+\b', '', sent)
    
    return sent

Function for preprocessing tweets:

def preprocess_tweets(tweets_df):
    from spellchecker import SpellChecker
    spell = SpellChecker(distance=1)
    
    tweets_df["textPreprocessed"] = tweets_df.text
    tweets_df["textPreprocessed"] =  tweets_df["textPreprocessed"].str.replace("@\w+", "") # remove AT's
    tweets_df["textPreprocessed"] = tweets_df["textPreprocessed"].str.replace("^(RT)+", "") # Remove RT at beginning of retweets
    
    # Add stuff probably
    
    tweets_df["textPreprocessed"] = tweets_df["textPreprocessed"].apply(preprocess)
    tweets_df["textPreprocessed"] = tweets_df["textPreprocessed"].apply(
        lambda x : " ".join([spell.correction(el) for el in x.split()]))

Applying this function to the datasets:

preprocess_tweets(CESTweets)
CESTweets.head()

preprocess_tweets(SMTweets)
SMTweets.head()

preprocess_tweets(SANTweets)
SANTweets.head()

## Sentiwordnet Sentiment Classifier

import nltk
from nltk.stem import WordNetLemmatizer, PorterStemmer
from nltk.corpus import sentiwordnet as swn
nltk.download('universal_tagset')
nltk.download('sentiwordnet')

The SWNClassifier takes in input the pre-processed tweets and  works as follows:


- Each tweet is split in tokens using nltk's tokenizer;
- A *Part of Speech* tag is assigned to each token with nltk's `pos_tag` function;
- Each tag is translated to a SentiWordNet tag;
- Each token/tag pair is assigned the positivity and negativity score defined by SentiWordNet;
- The positivity and negativity score of a tweet is computed as the sum of positivity and negativity scores of the constituting tokens;
- If both positivity and negativity scores are 0, the tweet is labelled as neutral. If the positivity score is greater than the negativity score, the tweet is labelled as positive, and negative otherwise.

The function returns, for a list of tweets:

- Their tokens and their tags
- Their positivity score
- Their negativity score
- Their sentiment score (-1 for negative, 0 for neutral, 1 for positive)

def SWNClassifier(X):
    # Adapted from https://towardsdatascience.com/sentiment-analysis-on-swachh-bharat-using-twitter-216369cfa534
    lem = WordNetLemmatizer()
    pstem = PorterStemmer()
    X_tagged = []
    li_swn=[]
    li_swn_pos=[]
    li_swn_neg=[]
    missing_words=[]
    for i in range(len(X)):
        text = X[i]
        tokens = nltk.word_tokenize(text)
        tagged_sent = nltk.tag.pos_tag(tokens)
        store_it = [(word, nltk.tag.map_tag('en-ptb', 'universal', tag)) for word, tag in tagged_sent]
        X_tagged.append(store_it)
        #print("Tagged Parts of Speech:",store_it)

        pos_total=0
        neg_total=0
        for word,tag in store_it:
            # print(tag)
            if(tag=='NOUN'):
                tag='n'
            elif(tag=='VERB'):
                tag='v'
            elif(tag=='ADJ'):
                tag='a'
            elif(tag=='ADV'):
                tag = 'r'
            else:
                tag='nothing'

                
            if(tag!='nothing'):
                concat = word+'.'+tag+'.01'
                try:
                    this_word_pos=swn.senti_synset(concat).pos_score()
                    this_word_neg=swn.senti_synset(concat).neg_score()
                    # print(word,tag,':',this_word_pos,this_word_neg)
                except Exception as e:
                    wor = lem.lemmatize(word)
                    concat = wor+'.'+tag+'.01'
                    # Checking if there's a possiblity of lemmatized word be accepted into SWN corpus
                    try:
                        this_word_pos=swn.senti_synset(concat).pos_score()
                        this_word_neg=swn.senti_synset(concat).neg_score()
                    except Exception as e:
                        wor = pstem.stem(word)
                        concat = wor+'.'+tag+'.01'
                        # Checking if there's a possiblity of lemmatized word be accepted
                        try:
                            this_word_pos=swn.senti_synset(concat).pos_score()
                            this_word_neg=swn.senti_synset(concat).neg_score()
                        except:
                            missing_words.append(word)
                            continue
                pos_total+=this_word_pos
                neg_total+=this_word_neg
        li_swn_pos.append(pos_total)
        li_swn_neg.append(neg_total)

        if(pos_total!=0 or neg_total!=0):
            if(pos_total>neg_total):
                li_swn.append(1)
            else:
                li_swn.append(-1)
        else:
            li_swn.append(0)
            
    return X_tagged, li_swn_pos, li_swn_neg, li_swn

def scores(y_test, predictions):

    prec = precision_score(y_test, predictions) # Precision
    rec = recall_score(y_test, predictions) # Recall
    f1 = f1_score(y_test, predictions) # F1
    f2 = fbeta_score(y_test, predictions, 2) # F2
    cm = confusion_matrix(y_test, predictions)
    
    scores_strings = ["Test Precision",
                      "Test Recall", "F1", "F2"]
    
    scores = [prec, rec, f1, f2]
    
    print(("{:20s} {:.5f}\n"*4)[:-1].format(*itertools.chain(*zip(scores_strings, scores))))
    
    print(classification_report(y_test, predictions, digits=5))


### Evaluating the SWNClassifier

#### Evaluating the SWNClassifier on SanTweets

Let us compare the positive/negative/neutral labels assigned by the SWNClassifier with the original labels assigned to SANTweets.

%%time
SANTweets_tagged, SANTweets_SWN_POS, SANTweets_SWN_NEG, SANTweets_SWN_SENT = SWNClassifier(SANTweets.textPreprocessed.values)

SANTweets_SENT = SANTweets.sent.values

Evaluation on positive/negative/neutral labels:

print(classification_report(SANTweets_SENT, SANTweets_SWN_SENT, digits=5))

Evaluation on Neutral/Sentiment labels:

scores(SANTweets_SENT==0, np.array(SANTweets_SWN_SENT)==0)

#### Evaluating the SWNClassifier on SMTweets

Let us compare the positive/negative/neutral labels assigned by the SWNClassifier with the manual labels assigned to SMTweets.

%%time
SMTweets_tagged, SMTweets_SWN_POS, SMTweets_SWN_NEG, SMTweets_SWN_SENT = SWNClassifier(SMTweets.textPreprocessed.values)

SMTweets_MANUAL_SENT = SMTweets.sent.values

Evaluation on positive/negative/neutral labels:

print(classification_report(SMTweets_MANUAL_SENT, np.array(SMTweets_SWN_SENT), digits=5))

Evaluation on Neutral/Sentiment labels:

scores(SMTweets_MANUAL_SENT == 0, np.array(SMTweets_SWN_SENT)==0)

# Evaluating Multinomial Naive Bayes Classifier from previous notebook

Load joblib files:

import zipfile
with zipfile.ZipFile('joblib_data/tfidf_vect_nostemmer.zip', 'r') as zip_ref:
    zip_ref.extractall('joblib_data')

vectorizer = load('joblib_data/tfidf_vect_nostemmer.joblib')

clf = load('joblib_data/clf_nb_nostemmer.joblib')

## Evaluating MNB Classifier on Samsung Mobile Tweet Replies

Now we evaluate our classifier on the SMTweets. Because our classifier only outputs positive/negative, we have to filter out neutral tweets. Hence, we take into account:

- SMTweets manually labelled as positive/negative;
- SMTweets labelled as positive/negative by the SWNClassifier.

#### SMTweets manually labelled as positive or negative:

SMTweets_MANUAL_POS_NEG = SMTweets[SMTweets.sent != 0]

SMTweets_MANUAL_POS_NEG_x = SMTweets_MANUAL_POS_NEG.textPreprocessed.values
SMTweets_MANUAL_POS_NEG_y = SMTweets_MANUAL_POS_NEG.sent.values == 1

Manual labels: 139 negative, 19 positive

np.unique(SMTweets_MANUAL_POS_NEG_y, return_counts=True)

Applying the vectorizer:

SMTweets_MANUAL_POS_NEG_x_vect = vectorizer.transform(SMTweets_MANUAL_POS_NEG_x)
SMTweets_MANUAL_POS_NEG_x_vect

Predictions from the MNB Classifier:

%%time
predictions = clf.predict(SMTweets_MANUAL_POS_NEG_x_vect)

np.unique(predictions, return_counts = True)

MNB predictions against the manual labels:

scores(SMTweets_MANUAL_POS_NEG_y, predictions)

#### SMTweets labelled as positive or negative by the SWNClassifier:

SMTweets_SWN_SENT = np.array(SMTweets_SWN_SENT)
SMTweets_SWN_POS_NEG_x = SMTweets.textPreprocessed.values[SMTweets_SWN_SENT != 0]
SMTweets_SWN_POS_NEG_y = SMTweets_SWN_SENT[SMTweets_SWN_SENT != 0] == 1

SWN Labels: 134 negative, 127 positive.

np.unique(SMTweets_SWN_POS_NEG_y, return_counts=True)

Applying the vectorizer:

SMTweets_SWN_POS_NEG_x_vect = vectorizer.transform(SMTweets_SWN_POS_NEG_x)
SMTweets_SWN_POS_NEG_x_vect

Predictions from the MNB Classifier:

%%time
predictions = clf.predict(SMTweets_SWN_POS_NEG_x_vect)

np.unique(predictions, return_counts = True)

MNB predictions against the SWN labels on SMTweets:

scores(SMTweets_SWN_POS_NEG_y, predictions)

## Evaluating MNB Classifier on Sanders Analytics Tweets

Now we evaluate our classifier on the SANTweets. Because our classifier only outputs positive/negative, we have to filter out neutral tweets. Hence, we take into account:

- SANTweets originally labelled as positive/negative;
- SANTweets labelled as positive/negative by the SWNClassifier.

#### SANTweets labelled as positive or negative:

SANTweets_SENT = SANTweets.sent.values
SANTweets_POS_NEG_x = SANTweets.textPreprocessed.values[SANTweets_SENT != 0]
SANTweets_POS_NEG_y = SANTweets_SENT[SANTweets_SENT != 0] == 1

Original labels: 572 negative, 519 positive

np.unique(SANTweets_POS_NEG_y, return_counts=True)

Applying the vectorizer:

SANTweets_POS_NEG_x_vect = vectorizer.transform(SANTweets_POS_NEG_x)
SANTweets_POS_NEG_x_vect

Predictions from the MNB Classifier:

%%time
predictions = clf.predict(SANTweets_POS_NEG_x_vect)

np.unique(predictions, return_counts = True)

MNB predictions against the original SANTweets labels:

scores(SANTweets_POS_NEG_y, predictions)

#### SANTweets labelled as positive or negative by the SWNClassifier:

SANTweets_SWN_SENT = np.array(SANTweets_SWN_SENT)
SANTweets_SWN_POS_NEG_x = SANTweets.textPreprocessed.values[SANTweets_SWN_SENT != 0]
SANTweets_SWN_POS_NEG_y = SANTweets_SWN_SENT[SANTweets_SWN_SENT != 0] == 1

SWN SANTweets Labels: 1326 negative, 2167 positive.

np.unique(SANTweets_SWN_POS_NEG_y, return_counts=True)

Applying the vectorizer:

SANTweets_SWN_POS_NEG_x_vect = vectorizer.transform(SANTweets_SWN_POS_NEG_x)
SANTweets_SWN_POS_NEG_x_vect

Predictions from the MNB Classifier:

%%time
predictions = clf.predict(SANTweets_SWN_POS_NEG_x_vect)

np.unique(predictions, return_counts = True)

MNB predictions against the SWN labels on SANTweets:

scores(SANTweets_SWN_POS_NEG_y, predictions)

## Evaluating MNB Classifier on CESTweets

Because CESTweets do not have any label, we can only rely on the SWNClassifier.

CESTweets_tagged, CESTweets_SWN_POS, CESTweets_SWN_NEG, CESTweets_SWN_SENT = SWNClassifier(CESTweets.textPreprocessed.values)

#### CESTweets labelled as positive or negative by the SWNClassifier:

CESTweets_SWN_SENT = np.array(CESTweets_SWN_SENT)
CESTweets_SWN_POS_NEG_x = CESTweets.textPreprocessed.values[CESTweets_SWN_SENT != 0]
CESTweets_SWN_POS_NEG_y = CESTweets_SWN_SENT[CESTweets_SWN_SENT != 0] == 1

SWN CESTweets Labels: 465 negative, 1229 positive.

np.unique(CESTweets_SWN_POS_NEG_y, return_counts=True)

Applying the vectorizer:

CESTweets_SWN_POS_NEG_x_vect = vectorizer.transform(CESTweets_SWN_POS_NEG_x)
CESTweets_SWN_POS_NEG_x_vect

Predictions from the MNB Classifier:

%%time
predictions = clf.predict(CESTweets_SWN_POS_NEG_x_vect)

np.unique(predictions, return_counts = True)

MNB predictions against the SWN labels on SMTweets:

scores(CESTweets_SWN_POS_NEG_y, predictions)
